backend/requirement.txt
tensorflow
numpy
pillow
python-dotenv
fastapi